fix: add `.v0` prefix to `github-release` binary (#4)
This patch updates the `github-release` binary name. Originally, we were going to pull the software directly from GitHub (via `go get https://github.com/aktau/github-release`), but to avoid dealing with breaking changes, we're always installing v0 instead. As a side-effect of pulling the v0 binary down, we must reference it with a `.v0` suffix.