Add `docker-push-to-ecr`
This patch adds a script for building/pushing a Docker image to an ECR. As result of the "infra catchup" discussion yesterday, I decided to build a more general-purpose script to use in all of our microservices for pushing to their ECRs.

The script was based on the following:

- https://github.com/dequelabs/walnut/blob/develop/scripts/docker-push.sh
- https://github.com/dequelabs/axe-linter/blob/develop/.circleci/deploy-docker-image.sh
- https://github.com/dequelabs/axe-pro-ml-service/blob/develop/.circleci/deploy-docker-image.sh
- https://github.com/dequelabs/auth-service/blob/develop/docker/deploy.sh
- https://github.com/dequelabs/axe-netlify/blob/develop/.circleci/deploy-docker-image.sh
- https://github.com/dequelabs/attest-usage-service/blob/develop/.circleci/deploy-service.sh

We should be able to replace the above (and others service's custom build stuff) with this script.