feat(docker-push-to-ecr): Support pushing directly to production ECRs
This patch enables us to move away from the `-production` image hack. With it, we'll no longer push production images to our dev ECRs.

For now, this is opt-in behavior. To enable this, simply set the following environment variables prior to running the `docker-push-to-ecr` script:

- `PROD_ECR` - the production ECR address
- `PROD_AWS_SECRET_ACCESS_KEY` - secret access key for an IAM user with access to push to ECR
- `PROD_AWS_ACCESS_KEY_ID` - access key id for an IAM user with access to push to ECR

I will be testing this on a few of the micro-services I maintain over the next few weeks/sprints. Once I am confident that it's working as expected, I'll post in Deque's `#infrastructure` slack channel, letting everyone know to add the new variables to their CI setups. About a week after that, I will _remove_ the `-production` hack from this script.

See https://github.com/dequelabs/infrastructure/issues/231