feat(node): support gha env variables (#45)
# Background

In moving the `axe-core` deployment to GHA, it was discovered that this repository of scripts has no support for GHA env variables. While we _could_ re-implement the release scripting fresh for GHA, this would take some time to do and ensure alignment with the current scripts behavior.

This patch does 4 things:

- Adds support for the GitHub environment variables when CircleCI ones are not found
- Notes in the readme about this introduction of support starting, but not being complete. Thus not fully documented on behavior patterns yet.
- Pins the checkout workflow instead of using a loose tag (also updates it to current in the process.)
- Restricts the permissions of the workflow to none.

The patch here is non-breaking. The `GITHUB_` variables added should not exist in CircleCI environments. Even if they did, CircleCI takes priority.